implement dream team overall by coefficient

+++ LA_LIGA_REKREATIVO/Shared/Dto/PlayerStatsDto.cs
@@ -6,7 +6,7 @@
         //public int Season { get; set; }
         //General
         public int TotalMatches { get; set; }
-        public int TotalPoints { get; set; }
+        public double TotalPoints { get; set; }
         public int Wins { get; set; }
         public int Losts { get; set; }
         public int Draws { get; set; }
